Tamayo
Had to relink all the pages to the master page. If you're working on a
page please copy the work to another spot before syncing and then
reapply as needed. I also added the siteMap to every page so that we
have a that. Still working on the other navigation to each page and
trying to make it easy to drag and drop.