Admin page added
Added the Admin page which allows admin to modify/update username/email.

Things that don't work/need to be fixed:
1) Web.config in Admin is currently commented out as I haven't been able
to get it to work properly. Hopefully somebody can help here.
2) Cancel button does not currently work properly. It's only a redirect
and it saves the updating of roles.

Idea:
1) Maybe we should add a content template to myAccount page that shows
an extra link to administration page only if the user is an admin and
show regular content otherwise (This would cover requirement 8, which
I'm not sure if we have covered or not yet). Otherwise, I'm not too sure
how we want to handle accessing the actual admin page.