Added Theme Selection
I had to clean up the Themes area and rename them. Everything still
works. The "Dark" theme can be applied by the user to the accounts page,
but it still needs to be modified so that it saves in a cookie.

The scoreboard page was giving errors so i commented it out. Maybe
somebody can find the fix to that as well.

We need Antonio B to upload the images he used to the Images folder.